requirements.txt
ultralytics
opencv-python
supervision
pyyaml
